+++ Content.Server/_RMC14/LinkAccount/LinkAccountSystem.cs
@@ -1,4 +1,4 @@
-﻿using Content.Server._RMC14.Rules;
+﻿using Content.Server._RMC14.Rules.DistressSignal;
 using Content.Server.Administration.Logs;
 using Content.Server.Database;
 using Content.Server.GameTicking;
