+++ Content.Server/_RMC14/MapInsert/MapInsertSystem.cs
@@ -1,10 +1,9 @@
 ﻿using System.Numerics;
-using Content.Server._RMC14.Rules;
+using Content.Server._RMC14.Rules.DistressSignal;
 using Content.Server.Decals;
 using Content.Shared._RMC14.Areas;
 using Content.Shared._RMC14.Rules;
 using Content.Shared.Decals;
-using Content.Shared.GameTicking;
 using Robust.Server.Physics;
 using Robust.Shared.EntitySerialization.Systems;
 using Robust.Shared.Map;
