+++ Content.Server/_RMC14/Rules/RMCDistressSignalPrefixesComponent.cs
@@ -1,4 +1,5 @@
-﻿using Robust.Shared.Prototypes;
+﻿using Content.Server._RMC14.Rules.DistressSignal;
+using Robust.Shared.Prototypes;
 
 namespace Content.Server._RMC14.Rules;
