+++ Content.Server/_RMC14/Rules/SetOperationNameCommand.cs
@@ -1,11 +1,8 @@
-using Content.Server._RMC14.Rules;
+using Content.Server._RMC14.Rules.DistressSignal;
 using Content.Server.Administration;
 using Content.Server.GameTicking;
-using Content.Shared._RMC14.Rules;
 using Content.Shared.Administration;
-using Robust.Shared.Prototypes;
 using Robust.Shared.Toolshed;
-using Robust.Shared.Utility;
 
 namespace Content.Server._RMC14.Rules;
