+++ Resources/Locale/en-US/_RMC14/game-presets/distress-signal-preset.ftl
@@ -31,3 +31,10 @@ cm-distress-signal-jellies = Royal Jelly Awards:
 rmc-distress-signal-got-medal = {$receiver} is awarded the {$award}: '{$awardDescription}' by {$giver}
 
 rmc-distress-signal-got-jelly = {$receiver} is awarded the {$award}: '{$awardDescription}' by {$giver}
+
+rmc-distress-signal-ares-online = ARES. Online. Good morning, marines.
+rmc-distress-signal-bioscan-complete = Bioscan complete. No unknown lifeform signature detected.
+rmc-distress-signal-saving-report = Saving operational report to archive.
+rmc-distress-signal-final-scan = Commencing final systems scan in 3 minutes.
+rmc-distress-signal-admin-start-fail = Can't start distress signal. Requires at least {$minimum} xeno player but we have {$candidates}.
+rmc-distress-signal-operation-name = Operation {$name}
